0.1 Coming soon

+++ v.0.1/UBLS-server.py
@@ -15,13 +15,20 @@ clients = []
 dirname = os.path.dirname(__file__) + "/"
 serverSocket.listen(1)
 MAX_CLIENTS = 10
-a = input("(load <name> / newWorld) >>> ")
-if a[0:5].lower() == "load ":
-	if not a[-5:] == ".json": a+=".json"
-	with open(dirname+"server/saves/"+a[5:], "r") as f:
-		serverBlocks = json.load(f)
-else:
-	serverBlocks = []
+while True:
+	a = input("[INPUT] (load <name> / newWorld) >>> ")
+	if a[0:5].lower() == "load ":
+		if not a[-5:] == ".json": a+=".json"
+		if os.path.exists(dirname+"server/saves/"+a[5:]):
+			with open(dirname+"server/saves/"+a[5:], "r") as f:
+				serverBlocks = json.load(f)
+				print("[LOG]", "Successfully loaded world "+a[5:])
+				break
+		else:
+			print("[ERROR] {} does not exist".format(a[5:]))
+	else:
+		serverBlocks = []
+		break
 
 
 class Client:
@@ -52,12 +59,6 @@ class Client:
 				for i in data["deletedBlocks"]:
 					for bl in serverBlocks:
 						if i["pos"] == bl["pos"] and i["level"] == bl["level"]:
-							for block, inID in bl["connections"]:
-								for index, bl2 in enumerate(serverBlocks):
-									if block["pos"] == bl2["pos"] and block["level"] == bl2["level"]:
-										bl2["inputs"][inID] = False
-										serverBlocks[index] = updateBlock(bl2)
-
 							serverBlocks.remove(bl)
 
 				for i in data["updatedBlocks"]:
@@ -65,7 +66,6 @@ class Client:
 						if i["pos"] == bl2["pos"] and i["level"] == bl2["level"]:
 							serverBlocks[index] = i
 							serverBlocks[index] = updateBlock(serverBlocks[index])
-							print(i)
 
 				reply = {"newBlocks": [], "updatedBlocks": [], "deletedBlocks": []}
 				if len(self.oldBlocks) != len(serverBlocks):
@@ -102,31 +102,42 @@ class Client:
 
 
 def updateBlock(block):
-	if block["type"] == "Generator":
+	if block["type"] == "Generator" or block["type"] == "Lever":
 		for bl, inID in block["connections"]:
 			for index, bl2 in enumerate(serverBlocks):
-				if bl2["pos"] == bl["pos"] and bl2["level"] == bl["pos"]:
+				if bl2["pos"] == bl["pos"] and bl2["level"] == bl["level"]:
 					bl2["inputs"][inID] = True
-					print(bl2["inputs"][inID])
 					serverBlocks[index] = updateBlock(bl2)
 	elif block["type"] == "Not":
-		if block["is_connected"][0]:
-			found = False
-			for bl in serverBlocks:
-				if bl["pos"] == block["is_connected"][1]["pos"] and bl["level"] == block["is_connected"][1]["level"]:
-					found = True
-					block["inputs"][0] = bl["out"]
-			# if not found:
-			# 	block["is_connected"] = [False]
-			# 	block["inputs"][0] = False
+		for index, b in enumerate(block["is_connected"]):
+			if b[0]:
+				for bl in serverBlocks:
+					if bl["pos"] == b[1]["pos"] and bl["level"] == b[1]["level"]:
+						block["inputs"][b[2]] = bl["out"]
+			elif block["inputs"][index] == True:
+				block["inputs"][index] = False
+
 		block["out"] = not block["inputs"][0]
 		for bl, inID in block["connections"]:
 			for index, bl2 in enumerate(serverBlocks):
-				if bl2["pos"] == bl["pos"] and bl2["level"] == bl["pos"]:
-					bl["inputs"][inID] = block["out"]
-					serverBlocks[index] = updateBlock(bl)
+				if bl2["pos"] == bl["pos"] and bl2["level"] == bl["level"]:
+					bl2["inputs"][inID] = block["out"]
+					serverBlocks[index] = updateBlock(bl2)
+	elif block["type"] == "OR":
+		for index, b in enumerate(block["is_connected"]):
+			if b[0]:
+				for bl in serverBlocks:
+					if bl["pos"] == b[1]["pos"] and bl["level"] == b[1]["level"]:
+						block["inputs"][b[2]] = bl["out"]
+			elif block["inputs"][index] == True:
+				block["inputs"][index] = False
+		block["out"] = block["inputs"][0] or block["inputs"][1] or block["inputs"][2]
+		for bl, inID in block["connections"]:
+			for index, bl2 in enumerate(serverBlocks):
+				if bl2["pos"] == bl["pos"] and bl2["level"] == bl["level"]:
+					bl2["inputs"][inID] = block["out"]
+					serverBlocks[index] = updateBlock(bl2)
 	for c in clients:
-		print(block)
 		c.updatedBlocks.append(block)
 	return block
 
@@ -139,8 +150,16 @@ def globalUpdateCycle():
 			if not a[-5:] == ".json": a+=".json"
 			if not os.path.exists(dirname+"server"): os.mkdir(dirname+"server")
 			if not os.path.exists(dirname+"server/saves"): os.mkdir(dirname+"server/saves")
-			with open(dirname+"server/saves/"+a[10:], "w") as f:
-				json.dump(serverBlocks, f)
+			if not os.path.exists(dirname+"server/saves/"+a[10:]):
+				with open(dirname+"server/saves/"+a[10:], "w") as f:
+					json.dump(serverBlocks, f)
+					print("[LOG]", "Successfully saved world "+a[10:])
+			else:
+				b = input("[INPUT] {} is already exists. Do you want to overwrite it? (y/n)".format(a[10:]))
+				if b == "y":
+					with open(dirname+"server/saves/"+a[10:], "w+") as f:
+						json.dump(serverBlocks, f)
+						print("[LOG]", "Successfully overwrited world "+a[10:])
 
 running = True
 globalUpdateCycleThread = threading.Thread(target=globalUpdateCycle)
